bot.pl: events - read from the events file
As they're in arrays, we can just use random access to them.  Remember
to reread on rehash. All events are in arrays under a hash indexed by
the letters 'W'/'L' for winning/losing, and 'G'/'C' for
godsend/catastrophe.

Move the built-in hog events to the events file. Also template them so
that the player's name can be substituted in arbitrarily many times.
Simplify and make the hog() function more data driven.

Signed-off-by: Phil Carmody <fatphil@asdf.org>