bot.pl: item levels - support lists, including counters
Best explained with an example. In items.txt, include names for the levels:

To sequence through linux 2.0, linux 2.1, linux 2.2, linux 2.3, linux 2.4,
linux 2.5, linux 2.6.0, linux 2.6.1, ..., then there are 2 options:

The dumb way:
 item1: n="OS"
 levels1: linux 2.0, linux 2.1, linux 2.2, linux 2.3, linux 2.4, linux 2.5, linux 2.6.0, linux 2.6.1, ...
(and those ... mean you have to type the whole thing out)

The smart way:
 item2: "OS"
 levels2: linux 2.%{0#+1#5}%, linux 2.6.%{0#+1}%

The %{start#delta#end}% notation may also have * rather than +/- as the
operation in the delta, but be warned that exponential growth is
dangerous.

Signed-off-by: Phil Carmody <fatphil@asdf.org>