bot.pl: change type 2 quest movement logic
Previously you'd head down the diagonal until you hit the right row
or column, and then you'd head in a straight line.

Now, you always move in the axis you're most distant on, and only
randomly move in the axis you're closer on, based on the ratio
of the two distances. Basically like a randomised Bresenham. The
path ends up mostly straight, but with a slight wobble. Much closer
to how you might cross terrain.

Signed-off-by: Phil Carmody <fatphil@asdf.org>