bot.pl: events - enable numeric macros in events
As the changes can be to a random extent, let the numeric part of the
message also be random. So you can now have events like:

  W Malcolm McLaren really likes %player%'s %{5##12}%" mohawk, so pushes %him%

Then a small push would correspond with 5", and a large push would be 12#.
The optional 3rd parameter species the step size, which defaults to 1 and
is necessary for fractional or large steps, or a range that decreases, e.g.:

  C has been given a grade %{5#-1#1}% crew-cut

Where grade 5 is minor, and grade 1 is the biggest calamity.

Item events currently aren't parametrised by a random number, they're
always 10% up or 10% down, so there's no need for a parameter. Make it
undef to trap any accidental erronious use of the macro in that context.

Signed-off-by: Phil Carmody <fatphil@asdf.org>