bot.pl: item_level told about indefinite articles
You don't want to find "a 6 place ratings", so have a special
marker that says "drop any leading article. That means item_level
needs to know if there's an article needed before the item level.
Add it as a parameter, and remove it if the level has a !a marker.

Note: this might not have any use at all, as I don't think I
can get simple counters to work without lots of work.

Signed-off-by: Phil Carmody <fatphil@asdf.org>