beer: HACK - first stab at items, events, and a config file
WILL NEED FIXING, the speed is way way too high in the conf files,
and the grammar is bollocksed up in a lot of the items.

The ability to understand the new style [ items ] save file is recent.
Make use of it by stubbing out the ones we can't get to make sense.

Signed-off-by: Phil Carmody <fatphil@asdf.org>