bot.pl: friendlier quit/restart messages
The admin can give a reason for the die/restart if he sees fit.

Also, the message doesn't need the admin's whole ident, only the nick,
as people can /whois on that if they really care (and if the admin is
somehow cloaked frmo whois, then we should respect that).

Signed-off-by: Phil Carmody <pc+dev@asdf.org>