bot.pl: add a saytype parameter to item_describe
And to all the helpers it calls too.

This is because we don't want to return a value to the caller to tell
it whether we said the type or not when expanding a unique's name,
so it should just tell us to append the type if it's sure it needs it.
item_describe's helpers can then decide if they've already said enough,
in which case they don't need to say the type, or if they haven't, in
which case they do appent the type.

Currently, nothing actually detects the types in the names, so there
should be no visible changes.

Signed-off-by: Phil Carmody <pc+dev@asdf.org>