bot.pl: pull article and saytype handling out of leaf functions
This permits item_describe to do all the logic for both uniques
and plain items. It also means the clients don't need to select
'a' or 'the' as the article.

Alas the full description of uniques, i.e. the "your enemies swoon..."
bit, is messing up the interface, and one client can't use the
item_describe interface, having to go straight to unique_describe
to get the "full" description. That's lame. The description should
be split.

Signed-off-by: Phil Carmody <pc+dev@asdf.org>