bot.pl: merge the two 1-on-1 fights into one function
They were clearly copy-pasted in the first place, as the collision
fight used to check whether you were colliding with the bot!

Parametrise them - the only differences are the relevance of
alignment on critical strikes (why? is that just oversight due
to unnecessary duplication of the code?) and the message:
"came upon" vs. "challenged". Only preserving the former due
to back-ass-ward compatibility.

Signed-off-by: Phil Carmody <pc+dev@asdf.org>