bot.pl: create a symmetrical fight helper
Presently, fights are asymmetrical - one player fights another, and
it's the instigator whose time changes. However, there's no "play to win"
or "play to avoid a loss" symmetry. This new helper takes a flag which
has the following values:
  -1: only time may be subtracted, from the winner.
   0: as before, instigator (player 0) has his time changed.
  +1: only time may be added, to the loser.

Of course, if the bot has to change time, which can only happen if the
flag's non-zero, or he's player 0 with flag=0, then the thing's a nonsense,
and in that case, just back away...

Presently no hooks call this.

Signed-off-by: Phil Carmody <pc+dev@asdf.org>