bot.pl: bugfix - the bot doesn't have a gender
Can't use the them() funcion on it, as it has no %rps entry.
This seems to be the only caller that can have $primnick as
a player name, so we can keep the core helpers simple and
naive.

This was a horrible bug, the them() function was causing an
%rps record to be created for thebot, which was causing deep
recursion in itemsum later.

Signed-off-by: Phil Carmody <pc+dev@asdf.org>