bot.pl: move players in random order
If the player hash doesn't change, the order in which players
move will be fixed. That means that pA will always get to a square
before pB for all A<B. So collision fights will always be the same
way round. This patch randomises the movement, so such things don't
happen.

Signed-off-by: Phil Carmody <pc+dev@asdf.org>