bot.pl: add penance to stop times spiralling out of control
With few players, an unlucky one with poor kit will keep losing fights,
and her time to level up will exponentially grow from the losses.

Prevent this spiralling out of control by pulling the player's
level down by one, and resetting their clock, so they have to
do the whole previous level again. They do not pass go, and do
not collect 200 pounds.

The message has 2 parts, first the supplication, which depends on
the player's alignment, PG, PN, and PE, and secondly the penance
itself, P. Defaults are provided for all of these messages if no
themed ones are available.

Note that the level down takes place between supplication and
pennance, so that the %level% macro expands to different numbers.

Signed-off-by: Phil Carmody <pc+dev@asdf.org>