bot.pl: permit restriction of some events to specific genders
An event with the prefix %!+chars% or %!-chars% will be rejected
if the (first) player's gender isn't (+) or is (-) in 'chars',
where chars can be any or all of 'm', 'f', and 'n'. A player with
'u' or 'pc' gender is considered able to undergo all events.

Intended only for C and G events initially, but this expansion
applies to pretty much all event types so could be used elsewhere.
(Note - item calamities/godsends are *not* C/G events.)

The syntax is crummy, and should be made more easily usable

Signed-off-by: Phil Carmody <pc+dev@asdf.org>