messagecheck interval
The bot now stores the last time it receives and/or sent a message. If the bot has not sent or received a message in longer than the messagecheck interval, assume the bot has lost connection to server and perform a reconnect. This provides much more robust disconnect detection than the old code provided